Sync conference power levels on join
When the user joins the conference, we catch the
`muc-occupant-joined` event and make them an owner
if they were marked as one during auth.

During auth, if we get power levels back from UVS
(see https://github.com/matrix-org/matrix-user-verification-service/pull/7), we compare the users power level in the
room to "state_default" of the room and if the
user is equal or more, we mark them as owner.
This allows them to do moderation in the Jitsi
conference side.